requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
Pillow==11.3.0
openpyxl==3.1.5
gunicorn==22.0.0
whitenoise==6.6.0